Move `Frame` to the `jit_ir` module.
This was previously in `trace_builder`, which is clearly the wrong
place, because `jit_ir::GuardInfo` refers to `Frame`. This commit
minimally moves the struct to the `jit_ir` module.
